+++ protocol-designer/src/steplist/actions/index.js
@@ -1,10 +1,4 @@
 // @flow
-import handleFormChange from './handleFormChange'
-
 export * from './actions'
 export * from './types'
 export * from './thunks'
-
-export {
-  handleFormChange,
-}
